frontend code refactoring
