+++ market-ui/src/Constants.ts
@@ -1,7 +1,2 @@
 export const MICROCCD_IN_CCD = 1000000;
 export const MARKET_CONTRACT_ADDRESS = { index: BigInt(984), subindex: BigInt(0) };
-export const TRANSFER_ENERGY = BigInt(6000);
-export const NFT_CONTRACT_MODULE_REF="8e8e7beaa6a1ced4041077c641efb10bbf0f6b8c4f3614d1861d06a6fc6b40dc";
-export const NFT_CONTRACT_SCHEMA = "ffff020100000008000000434953322d4e465400090000000900000062616c616e63654f6602100114000200000008000000746f6b656e5f69641d0007000000616464726573731502000000070000004163636f756e7401010000000b08000000436f6e747261637401010000000c10011b25000000040000006d696e7400140002000000050000006f776e65721502000000070000004163636f756e7401010000000b08000000436f6e747261637401010000000c06000000746f6b656e7312001d001400020000000300000075726c1601040000006861736816010a0000006f70657261746f724f66021001140002000000050000006f776e65721502000000070000004163636f756e7401010000000b08000000436f6e747261637401010000000c07000000616464726573731502000000070000004163636f756e7401010000000b08000000436f6e747261637401010000000c1001010f000000736574496d706c656d656e746f72730014000200000002000000696416000c000000696d706c656d656e746f727310020c08000000737570706f727473021001160010011503000000090000004e6f537570706f72740207000000537570706f72740209000000537570706f72744279010100000010000c0d000000746f6b656e4d657461646174610210011d0010011400020000000300000075726c160104000000686173681502000000040000004e6f6e650204000000536f6d650101000000132000000002080000007472616e7366657200100114000500000008000000746f6b656e5f69641d0006000000616d6f756e741b250000000400000066726f6d1502000000070000004163636f756e7401010000000b08000000436f6e747261637401010000000c02000000746f1502000000070000004163636f756e7401010000000b08000000436f6e747261637401020000000c160104000000646174611d010e0000007570646174654f70657261746f720010011400020000000600000075706461746515020000000600000052656d6f7665020300000041646402080000006f70657261746f721502000000070000004163636f756e7401010000000b08000000436f6e747261637401010000000c04000000766965770114000300000005000000737461746510020f1502000000070000004163636f756e7401010000000b08000000436f6e747261637401010000000c1400020000000c0000006f776e65645f746f6b656e7310021d00090000006f70657261746f727310021502000000070000004163636f756e7401010000000b08000000436f6e747261637401010000000c0a000000616c6c5f746f6b656e7310021d00080000006d6574616461746110020f1d001400020000000300000075726c160104000000686173681601"
-export const MARKET_CONTRACT_SCHEMA = "FFFF02010000000A0000004D61726B65742D4E465400030000000300000061646400140003000000140000006E66745F636F6E74726163745F616464726573730C08000000746F6B656E5F69641D000500000070726963650A040000006C69737401140101000000100114000300000008000000746F6B656E5F69641D0008000000636F6E74726163740C0500000070726963650A080000007472616E7366657200140003000000140000006E66745F636F6E74726163745F616464726573730C08000000746F6B656E5F69641D0002000000746F0B";
-export const MARKET_CONTRACT_SCHEMA_BASE64 = "//8CAQAAAAoAAABNYXJrZXQtTkZUAAMAAAADAAAAYWRkABQAAwAAABQAAABuZnRfY29udHJhY3RfYWRkcmVzcwwIAAAAdG9rZW5faWQdAAUAAABwcmljZQoEAAAAbGlzdAEUAQEAAAAQARQAAwAAAAgAAAB0b2tlbl9pZB0ACAAAAGNvbnRyYWN0DAUAAABwcmljZQoIAAAAdHJhbnNmZXIAFAADAAAAFAAAAG5mdF9jb250cmFjdF9hZGRyZXNzDAgAAAB0b2tlbl9pZB0AAgAAAHRvCw==";
