+++ market-ui/src/components/Cis2Init.tsx
@@ -1,0 +1,50 @@
+import { WalletApi } from "@concordium/browser-wallet-api-helpers";
+import { ContractAddress } from "@concordium/web-sdk";
+import { Paper, Typography, Button } from "@mui/material";
+import { useState } from "react";
+import { initContract } from "../models/Cis2Client";
+
+function Cis2Init(props: {
+	provider: WalletApi;
+	account: string;
+	onDone: (address: ContractAddress) => void;
+}) {
+	let [state, setState] = useState({
+		error: "",
+		processing: false,
+	});
+
+	function onOkClicked() {
+		setState({ ...state, processing: true });
+
+		initContract(props.provider, props.account)
+			.then((address) => {
+				setState({ ...state, processing: false });
+				props.onDone(address);
+			})
+			.catch((err: Error) => {
+				setState({ ...state, processing: false, error: err.message });
+			});
+	}
+
+	return (
+		<Paper>
+			<h3>Deploy NFT Collection</h3>
+			<form>
+				<div>{state.error && <Typography>{state.error}</Typography>}</div>
+				<div>{state.processing && <Typography>Deploying..</Typography>}</div>
+				<div>
+					<Button
+						variant="contained"
+						disabled={state.processing}
+						onClick={() => onOkClicked()}
+					>
+						Deploy
+					</Button>
+				</div>
+			</form>
+		</Paper>
+	);
+}
+
+export default Cis2Init;
