+++ market-ui/src/components/NftList.tsx
@@ -1,14 +1,16 @@
 import { WalletApi } from "@concordium/browser-wallet-api-helpers";
 import ImageList from "@mui/material/ImageList";
 import Container from "@mui/material/Container";
-import { TokenListItem } from "../models/TokenListItem";
-import NftListItem from "./NftListItem";
 import { ContractAddress } from "@concordium/web-sdk";
+
+import NftListItem from "./NftListItem";
+import { TokenListItem } from "../models/MarketplaceTypes";
 
 function NftList(props: {
 	marketContractAddress: ContractAddress;
 	tokens: TokenListItem[];
 	provider: WalletApi;
+	account: string;
 }) {
 	return (
 		<Container>
@@ -16,9 +18,10 @@ function NftList(props: {
 				{props.tokens.map((t) => (
 					<NftListItem
 						provider={props.provider}
+						account={props.account}
 						marketContractAddress={props.marketContractAddress}
 						item={t}
-						key={t.tokenId + t.contract.index + t.contract.subIndex}
+						key={t.tokenId + t.contract.index + t.contract.subindex}
 					/>
 				))}
 			</ImageList>
