+++ market-ui/src/components/NftListItem.tsx
@@ -1,27 +1,29 @@
-import { MouseEventHandler, useEffect, useState } from "react";
-import { WalletApi } from "@concordium/browser-wallet-api-helpers";
+import { useEffect, useState } from "react";
 import ImageListItem from "@mui/material/ImageListItem";
 import ImageListItemBar from "@mui/material/ImageListItemBar";
 import IconButton from "@mui/material/IconButton";
 import ShoppingCartIcon from "@mui/icons-material/ShoppingCart";
 import CheckIcon from "@mui/icons-material/Check";
+import { WalletApi } from "@concordium/browser-wallet-api-helpers";
+import { ContractAddress } from "@concordium/web-sdk";
 
-import { TokenListItem } from "../models/TokenListItem";
-import { Metadata } from "../models/Types";
-import { MarketplaceClient } from "../models/MarketPlaceContractClient";
 import {
 	ensureValidOutcome,
-	loadNft,
+	fetchJson,
 	toLocalstorageKey,
 } from "../models/Utils";
-import { ContractAddress } from "@concordium/web-sdk";
+import { TokenListItem } from "../models/MarketplaceTypes";
+import { getTokenMetadata } from "../models/Cis2Client";
+import { transfer } from "../models/MarketplaceClient";
+import { Metadata } from "../models/Cis2Types";
 
 function NftListItem(props: {
 	item: TokenListItem;
 	provider: WalletApi;
+	account: string;
 	marketContractAddress: ContractAddress;
 }) {
-	const { item, provider } = props;
+	const { item, provider, account, marketContractAddress } = props;
 
 	let [state, setState] = useState({
 		isLoading: true,
@@ -33,7 +35,14 @@ function NftListItem(props: {
 	});
 
 	const buy = (item: TokenListItem) => {
-		MarketplaceClient.transfer(props.marketContractAddress, item, provider)
+		transfer(
+			provider,
+			account,
+			marketContractAddress,
+			item.contract,
+			item.tokenId,
+			item.price
+		)
 			.then((o) => ensureValidOutcome(o))
 			.then((_) => {
 				setState({
@@ -49,38 +58,35 @@ function NftListItem(props: {
 	};
 
 	useEffect(() => {
-		let nftJson = localStorage.getItem(toLocalstorageKey(item));
-		if (nftJson) {
-			let metadata: Metadata = JSON.parse(nftJson);
+		let setStateMetdata = (metadata: Metadata) =>
 			setState({
 				...state,
 				isLoading: false,
 				url: metadata.display.url,
-				// url: "https://source.unsplash.com/user/c_v_r/100x100",
-				// url: "https://picsum.photos/200/300",
 				name: metadata.name,
 				desc: metadata.description,
 				price: item.price,
 			});
+
+		let nftJson = localStorage.getItem(toLocalstorageKey(item));
+		if (nftJson) {
+			setStateMetdata(JSON.parse(nftJson));
 		} else {
-			loadNft(provider, item).then((metadata) => {
-				localStorage.setItem(toLocalstorageKey(item), JSON.stringify(metadata));
-				setState({
-					...state,
-					isLoading: false,
-					url: metadata.display.url,
-					// url: "https://picsum.photos/200/300",
-					name: metadata.name,
-					desc: metadata.description,
-					price: item.price,
+			getTokenMetadata(provider, account, item.contract, item.tokenId)
+				.then((m) => fetchJson<Metadata>(m.url))
+				.then((metadata) => {
+					localStorage.setItem(
+						toLocalstorageKey(item),
+						JSON.stringify(metadata)
+					);
+					setStateMetdata(metadata);
 				});
-			});
 		}
 	}, [item]);
 
 	return (
 		<ImageListItem
-			key={item.tokenId + item.contract.index + item.contract.subIndex}
+			key={item.tokenId + item.contract.index + item.contract.subindex}
 		>
 			<img src={state.url} srcSet={state.url} alt={state.name} loading="lazy" />
 			<ImageListItemBar
