+++ market-ui/src/models/MarketplaceTypes.ts
@@ -1,0 +1,11 @@
+import { ContractAddress } from "@concordium/web-sdk";
+
+export type TokenList = TokenListItem[];
+export interface TokenListItem {
+	/**
+	 * Hex of token Id
+	 */
+	tokenId: string;
+	contract: ContractAddress;
+	price: bigint;
+}
