--- market-ui/src/models/Types.ts
@@ -1,16 +1,0 @@
-export interface Metadata {
-	name: string;
-	description: string;
-	display: Display;
-	attributes: Attribute[];
-}
-
-export interface Display {
-	url: string;
-}
-
-export interface Attribute {
-	name: string;
-	type: string;
-	value: string;
-}
